Change implementation inheritance to sealed class
